requirements.txt
# Core
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1

# Database
psycopg2-binary==2.9.9

# Email
# (smtplib is built into Python — no install needed)

# Web server (feedback loop)
flask

# Dashboard
streamlit==1.35.0
pandas>=2.0.0

# Airflow
apache-airflow==2.9.2

# Utilities
lxml==5.2.2
